Simplify the input logic
People almost never press more than one hotkey for the same move at the
same time (for example, D + RIGHT). However, the code for handling that
was adding a significant amount of complexity. Let's just get rid of it.
As a bonus, the new implementation is closer to the old implementation
so you could also say it's more "authentic".