Render the text dynamically
Instead of creating a sprite sheet for all our text, have a sprite sheer
for the font and render the text dynamically. This will make it easier
to update the text (including the version name) without needing to
recreate all the fonts. It also ensures that the various text borders
are consistent. It turns out that in the previous sprite sheet some of
the boxes were off by a pixel or two...