Use atexit to call SDL_Quit
The SDL documentation seems to strongly recommend calling SDL_Quit
before we exit the program, so to ensure we do that even in the case of
errors we tie it to atexit (as suggested by the SDL docs[1]).

Additionally, we got rid of all the calls to DestroyTexture,
DestroyRenderer and DestroyWindow that we had at the end of the program.
I expect that any resources that those are holding will get released
automatically by the operating system. Furthermore, we already weren't
calling these in the case of errors. So in a way, not calling them
before a successful exit is just being consistent with that.

[1] https://wiki.libsdl.org/SDL_Quit