Generate the config.h in the configure script
Since the XJUMP_VERSION and the XJUMP_DATADIR are hardcoded into the
compiled executable, there is no point in trying to do something like
`make prefix=blah install`, after the file was already compiled with
a different prefix. To that end, we're reorganizing things to make it
clear that some things can only be changed via the configure script.