Fix a bug when the installation directory does not exist
If we tried to use readlink -f to find the installation directory, that
would fail if the parent directory does not exist. That is because
readlink -f only allows the last component of the path to be missing.

The solution is to use readlink -m or readpath -m. The problem is that
from what I've heard, these aren't always installed or might not always
support the flags that we want. As a compromise, make it so that if the
path already starts with "/" then we don't try to canonicalize it.