Treat empty --prefix as default
The error message that we would get in this case was misleading because
it asked the user to provide an absolute path, even if such absolute
path is not actually needed.

We could add extra code to give a better error message to the empty
directory names but on second thought, who really needs that? Let's keep
things simple and just assume that if the directory name is empty then
we use the default value.