Embrace the dynamism in the GNU convention
Whenever we write something like `$(datadir)` or `$(CFLAGS)` in the
makefile, it will be recursively expanded by Make and likely go through
word expansion in the shell as well. Therefore, we shouldn't try to
pretend that these thing are regular strings that we need to properly
escape. We should assume that they will be expanded, and embrace it.